Implement dynamic protocol setting
Added functionality to dynamically set the protocol to either http or https based on the USE_HTTPS environment variable.

This change allows the application to switch between secure (https) and non-secure (http) modes based on the environment it's running in.

The protocol is used when constructing the NODE_SERVER_URL and FLASK_SERVER_URL.